+++ src/layout/layout_types.h
@@ -31,10 +31,10 @@ typedef struct window_list_t {
 } window_list_t;
 
 typedef struct workarea_t {
-  uint16_t x;
-  uint16_t y;
-  uint16_t w;
-  uint16_t h;
+  uint32_t x;
+  uint32_t y;
+  uint32_t w;
+  uint32_t h;
 } workarea_t;
 
 typedef struct grid_cell_t {
